\name{arf-example-data}
\docType{data}
\alias{.example}
\alias{arf-full-data}
\alias{makeExample}
\title{Activated Region Fitting Example Dataset}
\description{
  Simulated dataset for the ARF3DS4 package. Run \code{makeExample(path)} to create an ARF directory structure. The experiment is automatically loaded. 
}
\usage{makeExample(path)}
\arguments{
  \item{path}{Optional path to make the example experiment in. By default uses the path of the package installation.}
}
\format{An ARF experiment for one subject, and two conditions.}
\keyword{datasets}
