\name{arfversion}
\alias{arfversion}
\title{ Check the version of ARF objects.}
\description{
Print version information of an ARF object.
}
\usage{
arfversion(object)
}
\arguments{
  \item{object}{An arf object.}
 }
\value{
None. The function prints the current version of the object.
}
\seealso{
	\code{\linkS4class{version}}
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\keyword{ models }