\name{experiment-class}
\docType{class}
\alias{experiment-class}

\title{S4 experiment class}
\description{experiment}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("experiment", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the experiment (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
