\name{extr-repl-fmridata}
\alias{[,fmri.data-method}
\alias{[<-,fmri.data-method}
\title{Extract replace elements of an fmri.data object}
\description{Extract replace elements of an fmri.data object}
\details{Extract and replace use indexing as an array of dimensions set by the fmri.data object (3D and 4D)}
\value{Returns an object of class fmri.data (in the case of replacment) else returns the appropriate part of the array.}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{fmri.data}}
}
\keyword{ models }