\name{getExp}
\alias{getExp}

\title{ Access experiment variables.}
\description{
Access the experiment (if loaded).
}
\usage{
getExp()
}

\value{
Returns an object of class \code{\linkS4class{experiment}}.
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{experiment}}, \code{\link{loadExp}}
}
\keyword{ models }

