\name{loadOptions}
\Rdversion{1.1}
\alias{loadOptions}
\title{ Load arf3DS4 Options }
\description{
Load an "options" object belonging to an arf model.
}
\usage{
loadOptions(arfmodel)
}
\arguments{
  \item{arfmodel}{Object of class "model" (see \code{\linkS4Class{model}})}
}
\value{
Returns an object of class \code{\linkS4Class{options}}.
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{model}}, \code{\linkS4class{options}}
}
\keyword{ models }
