\name{makeResiduals}
\Rdversion{1.1}
\alias{makeResiduals}
\title{ Calculate model residuals }
\description{
makeResiduals calculates residuals of an arf model and writes them to a binary file.
}
\usage{
makeResiduals(arfmodel)
}
\arguments{
  \item{arfmodel}{An arf model object (see \code{\linkS4Class{model}}).}
}
\value{
Returns \code{TRUE} when successful.
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{model}}
}
\keyword{ models }
