\name{options-class}
\docType{class}
\alias{options-class}

\title{S4 options class}
\description{options}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("options", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the options (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
