\name{readData}
\Rdversion{1.1}
\alias{readData}
\title{read in a NIFTI/ANALYZE file.}
\description{readData reads in a NIFTI/ANALYZE file.}
\usage{
readData(filename)
}
\arguments{
  \item{filename}{Name of the file to be read in.}
}
\value{
Returns an object of class \code{\linkS4class{fmri.data}}.
}
\references{ See the Nifti homepage (\url{http://nifti.nimh.nih.gov/}) for more information on the nifti filetype.}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{fmri.data}}
}
\keyword{ models }
\keyword{ methods }
