\name{readDerivs}
\Rdversion{1.1}
\alias{readDerivs}
\title{Read derivative file}
\description{Read in derivatives}
\usage{
readDerivsarfmodel)
}
\arguments{
  \item{arfmodel}{A valid arf model object.}
}
\value{
Returns an object of class "matrix".
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{model}}, 
}
\keyword{ models }