\name{saveFunc}
\Rdversion{1.1}
\alias{saveFunc}
\title{Save functional object.}
\description{Save an object of class "functional".}
\usage{saveFunc(functional)}
\arguments{
  \item{functional}{Object of class \code{\linkS4class{functional}}}
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{functional}}, \code{\link{loadFunc}}
}
\keyword{ models }