\name{saveModel}
\Rdversion{1.1}
\alias{saveModel}
\title{Save a model object}
\description{
Save a model object to the model.Rda file.
}
\usage{
saveModel(arfmodel)
}
\arguments{
  \item{arfmodel}{An object of class \code{\linkS4class{model}}.}
}
\value{
None
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{model}}
}
\keyword{ models }
\keyword{ methods }
