\name{saveReg}
\Rdversion{1.1}
\alias{saveReg}
\title{Save registration object}
\description{Save registration object}
\usage{
saveReg(registration)
}
\arguments{
  \item{registration}{Object of class \code{\linkS4class{registration}}}
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{functional}}, \code{\linkS4class{experiment}}
}
\keyword{ models }
