\name{sequence-class}
\docType{class}
\alias{sequence-class}

\title{S4 sequence class}
\description{sequence}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("sequence", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the sequence (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
