\name{wald-class}
\docType{class}
\alias{wald-class}

\title{S4 wald class}
\description{wals}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("wald", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the wald (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
