\name{arf-example-data}
\docType{data}
\alias{makeExample}
\alias{.example}
\title{Activated Region Fitting Example Dataset}
\description{
  Simulated dataset for the ARF3DS4 package. Run \code{makeExample(path)} to create an ARF directory structure. The experiment is automatically loaded. 
}
\usage{makeExample(path)}
\format{An ARF experiment for one subject, and two conditions.}
\keyword{datasets}
