\name{arf-example-data}
\docType{data}
\alias{.example}
\alias{makeExample}

\title{Activated Region Fitting Example Dataset}
\description{
  Simulated dataset for the ARF3DS4 package. Run \code{makeExample(path)} to create an ARF directory structure in 'path'. The experiment is automatically loaded. 
}
\usage{makeExample(path)}
\arguments{
  \item{path}{Optional path to make the example experiment in. By default uses the path of the package installation.}
}
\format{An ARF experiment for one subject ('wouter'), and one condition ('A').}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\keyword{datasets}
