\name{arfversion}
\alias{arfversion}

\title{ Which version has an arf-object.}
\description{
Print the version of an arf-object.
}
\usage{
arfversion(object)
}
\arguments{
  \item{object}{An arf object}
 }
\value{
Prints the current version of the object.
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\keyword{ models }