\name{createRegs}
\Rdversion{1.1}
\alias{createRegs}
\title{ Create registration directories for runs of an ARF condition.
}
\description{
createRegs creates registration directories for runs of an ARF condition and sets the appropriate paths.
}
\usage{
createRegs(arfdata)
}
\arguments{
  \item{arfdata}{An object of class "data" (\code{\linkS4class{data}}) from a condition.}
}
\value{
None
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{data}}, \code{\linkS4class{registration}}, \code{\link{setRegFiles}}, \code{\link{setRegParams}}
}

\keyword{ methods }
