\name{data-class}
\docType{class}
\alias{data-class}

\title{S4 data class}
\description{data}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("data", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the data (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
