\name{loadExp}
\alias{loadExp}

\title{ load an ARF Experiment }
\description{
  Load an ARF experiment and set all appropriate objects within that experiment.
}
\usage{
loadExp(filename)
}

\arguments{
  \item{filename}{ Name of the experiment file }
}
\details{
  loadExp checks the experiment and changes the rootpaths of all objects to match the experiment,
}
\value{
 returns the experiment object
}

\author{ W.D.Weeda }

\keyword{ methods }
