\name{loadRda}
\alias{loadRda}

\title{ load a Rda file to an object }
\description{
  load Rdata files into an object (in stead of directly as an object to the workspace) 
}
\usage{
loadRda(file)
}

\arguments{
  \item{file}{ name of the Rda file }
}

\value{
  returns the loaded object
}

\author{ W.D. Weeda }

\keyword{ methods }

