\name{loadStart}
\Rdversion{1.1}
\alias{loadStart}
\title{ Load arf3DS4 starting values }
\description{
Load the starting values of an arf model from the .Rda file
}
\usage{
loadStart(arfmodel)
}
\arguments{
  \item{arfmodel}{Object of class "model" (see \code{\linkS4Class{model}})}
}
\value{
Returns a vector with starting values.
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{model}}
}
\keyword{ models }