\name{minBIC}
\alias{minBIC}
\title{Model minima BIC}
\description{Calculate and show the BIC, minimum and number of regions of all models in a condition.}
\usage{
minBIC(subject, condition)
}
\arguments{
  \item{subject}{Subject.}
  \item{condition}{Condition.}
}

\value{Returns a data.frame with information on all regions.}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{fmri.data}}
}
\keyword{ models }
