\name{model-class}
\docType{class}
\alias{model-class}

\title{S4 model class}
\description{model}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("model", ...)}.

}
\section{Slots}{
	 \describe{
    \item{\code{name}:}{Name of the model (\code{"character"})}
  }
}
\section{Methods}{
  \describe{
    \item{slots:}{Slots can be accessed via the standard accessor and replacment methods.}
	 }
}

\author{ Wouter D. Weeda }
\note{NOT YET IMPLEMENTED}
\keyword{classes}
