\name{partialCor}
\alias{partialCor}

\title{
Calculate partial correlations 
}
\description{
Calculate partial correlations of a correlation matrix.
}
\usage{
partialCor(R,n)
}
\arguments{
  \item{R}{An invertible correlation matrix.}
  \item{n}{Number of samples to obtain test-statistic.}
  
}

\value{
Returns a matrix of partial correlations
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\keyword{ models }