\name{saveReg}
\Rdversion{1.1}
\alias{saveReg}
\title{Save a registration object}
\description{Save a registration object}
\usage{
saveReg(registration)
}
\arguments{
  \item{registration}{Object of class \code{\linkS4class{registration}}.}
}
\author{ Wouter D. Weeda - w.d.weeda@gmail.com }
\seealso{
	\code{\linkS4class{registration}}, \code{\link{loadReg}}
}
\keyword{ models }
